detect: move grouping/building code into own file
Clean up main detect.c file by moving things related to rule
grouping out.
